% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{alter_legend}
\alias{alter_legend}
\title{Alter the Chart Legend}
\usage{
alter_legend(
  chart,
  ...,
  position = c("top", "left", "bottom", "right"),
  align = c("center", "start", "end"),
  display = TRUE,
  fullWidth = TRUE,
  onClick = NULL,
  onHover = NULL,
  onLeave = NULL,
  reverse = FALSE,
  labels = NULL,
  rtl = FALSE,
  textDirection = NULL
)
}
\arguments{
\item{chart}{A \code{chartjs} object.}

\item{...}{Additional named arguments to add to the legend.}

\item{position}{}

\item{align}{}

\item{display}{}

\item{fullWidth}{}

\item{onClick, onHover, onLeave}{}

\item{reverse}{}

\item{labels}{}

\item{rtl}{}

\item{textDirection}{}
}
\value{
A modified \code{chartjs} object.
}
\description{
See here for more detailed documentation:
\url{https://www.chartjs.org/docs/latest/configuration/legend.html}.
}
\examples{

}
