% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{alter_options}
\alias{alter_options}
\title{Alter Chart Options}
\usage{
alter_options(
  chart,
  ...,
  animDur = 400,
  hoverDelay = 0,
  resizeAnimDur = 0,
  tension = 0.4
)
}
\arguments{
\item{chart}{A \code{chartjs} object.}

\item{...}{Additional named arguments to add to the options.}

\item{animDur}{How long the initial plot building animation should take
(in milliseconds) (numeric scalar).}

\item{hoverDelay}{How long to wait after hovering before showing a tooltip
(in milliseconds) (numeric scalar).}

\item{resizeAnimDur}{How long the plot resizing animation should take
(in milliseconds) (numeric scalar).}

\item{tension}{The default tension to use for lines between the points
(0 for straight lines, 0.4 is default) (numeric scalar).}
}
\value{
A modified \code{chartjs} object.
}
\description{
See here for more detailed documentation:
  \url{https://www.chartjs.org/docs/latest/general/options.html}.
}
\examples{

}
