% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{axis_title}
\alias{axis_title}
\title{Modify an Axis Title}
\usage{
axis_title(
  chart,
  id,
  xOrY = tolower(substr(id, 1, 1)),
  ...,
  title,
  display = TRUE,
  lineHeight = 1.2,
  fontColor = "#666",
  fontSize = 12,
  fontStyle = "normal",
  fontFamily = "'Helvetica Neue', 'Helvetica', 'Arial', sans-serif",
  padding = 4
)
}
\arguments{
\item{chart}{A \code{chartjs} object.}

\item{id}{The ID of the axis to add a title to (character scalar).}

\item{xOrY}{Whether the axis is an x or y-axis (character scalar).}

\item{...}{Additional options to be used in the axis title (has no effect here).}

\item{title}{The title of the axis (character scalar).}

\item{display}{Whether to display the axis title (boolean scalar).}

\item{lineHeight}{The height of an individual line of text
(see \url{https://developer.mozilla.org/en-US/docs/Web/CSS/line-height}) (character/numeric scalar).}

\item{fontColor}{The color of the title font (character scalar).}

\item{fontSize}{The size of the title font (numeric scalar).}

\item{fontStyle}{Space-separated combination of values from
[normal, italic, oblique, initial, inherit] (character scalar).}

\item{fontFamily}{The font family of the title text (character scalar).}

\item{padding}{The padding to use around the axis title (numeric scalar).}
}
\value{
A modified \code{chartjs} object.
}
\description{
See: https://www.chartjs.org/docs/latest/axes/labelling.html
}
\examples{

}
