% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basechart.R
\name{baseChart}
\alias{baseChart}
\title{Base chart creation}
\usage{
baseChart(chartjs, type)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{type}{a character, the type of chart}
}
\description{
This is a low-level function called by the charts creation functions. It adds
colors and chart type to the object that will be interpreted by the Chart.js
library.
}
\keyword{internal}

