% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{chartjs_debug}
\alias{chartjs_debug}
\title{Diagnostic function for chartjs package}
\usage{
chartjs_debug()
}
\description{
This function helps diagnose issues with the chartjs package by checking
dependencies and creating a simple test chart.
}
