% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-proxy.R
\name{chartjs_proxy}
\alias{chartjs_proxy}
\title{Create a proxy object for Chart.js widget manipulation}
\usage{
chartjs_proxy(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{The output ID of the chart to control}

\item{session}{The Shiny session object (optional, defaults to current session)}
}
\value{
A chartjs_proxy object
}
\description{
This function creates a proxy object that can be used to update Chart.js
widgets that have already been rendered in a Shiny application, without
having to completely re-render the chart.
}
\examples{
\dontrun{
# In a Shiny server function
proxy <- chartjs_proxy("myChart")
proxy \%>\% chartjs_update_data(new_data)
}
}
