% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-proxy.R
\name{chartjs_remove_dataset}
\alias{chartjs_remove_dataset}
\title{Remove a dataset from the chart}
\usage{
chartjs_remove_dataset(proxy, index)
}
\arguments{
\item{proxy}{A chartjs_proxy object created with \code{\link{chartjs_proxy}}}

\item{index}{Index of the dataset to remove (0-based)}
}
\value{
The proxy object (for method chaining)
}
\description{
Remove a dataset from the chart
}
