% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-proxy.R
\name{chartjs_update_options}
\alias{chartjs_update_options}
\title{Update chart options via proxy}
\usage{
chartjs_update_options(proxy, options)
}
\arguments{
\item{proxy}{A chartjs_proxy object created with \code{\link{chartjs_proxy}}}

\item{options}{New options for the chart}
}
\value{
The proxy object (for method chaining)
}
\description{
Update chart options via proxy
}
