% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartjs.R
\name{chartjsOutput}
\alias{chartjsOutput}
\title{Widget output function for use in Shiny}
\usage{
chartjsOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

