% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{charts}
\alias{charts}
\alias{cjsBar}
\alias{cjsDoughnut}
\alias{cjsLine}
\alias{cjsPie}
\alias{cjsRadar}
\alias{polarAreaChart}
\alias{scatterChart}
\title{Draw charts}
\usage{
cjsBar(chartjs, stacked = FALSE)

cjsLine(chartjs)

cjsPie(chartjs)

cjsDoughnut(chartjs, cutout = 50)

cjsRadar(chartjs, colours = NULL)

polarAreaChart(chartjs, colours = NULL)

scatterChart(chartjs, colours = NULL)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{stacked}{a logical, defaults to FALSE. If TRUE, bars will be stacked
at each x value.}

\item{inner}{the optional percentage of the inner cutout for Doughnut
charts (defaults to 50)}
}
\description{
These functions are used to draw the different types of graphs.
}
\details{
For \code{colours}, the names of the list must be among those in \code{\link{colortypes}}.
The default colours are obtained from the \strong{Paired} palette from the
\code{\link[RColorBrewer]{RColorBrewer}} package.
}
\examples{
data <- list(mpg = mtcars$mpg, qsec = mtcars$qsec)

# Draw a bar chart
chartjs(data, labels = row.names(mtcars)) \%>\%
  cjsBar

# Draw a stacked bar chart
chartjs(data, labels = row.names(mtcars)) \%>\%
  cjsBar(stacked = TRUE)

# Draw a line chart
chartjs(data, labels = row.names(mtcars)) \%>\%
  cjsLine

# Draw a radar chart
chartjs(mtcars, mpg, qsec, labels = labels) \%>\%
  radarChart

# Draw a pie chart
chartjs(data = data$mpg[c(1:6)], labels = row.names(mtcars)[c(1:6)]) \%>\%
  cjsPie

# Draw a doughnut chart
chartjs(data = data$mpg[c(1:6)], labels = row.names(mtcars)[c(1:6)]) \%>\%
  cjsDoughnut(cutout = 80)

# Draw a scatter chart
chartjs(dataScatter) \%>\%
  scatterChart
}

