% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_series.R
\name{cjsAddSeries}
\alias{cjsAddSeries}
\alias{cjsAddSeries.cjs_bar}
\alias{cjsAddSeries.cjs_line}
\alias{cjsAddSeries.cjs_pie}
\alias{cjsAddSeries.cjs_polarArea}
\alias{cjsAddSeries.cjs_radar}
\title{Add a series to a chart}
\usage{
cjsAddSeries(x, ...)

\method{cjsAddSeries}{cjs_bar}(cjs, data, type = "bar", color = NULL,
  label = NULL, scale = NULL)

\method{cjsAddSeries}{cjs_line}(cjs, data, type = "line", color = NULL,
  label = NULL, scale = NULL)

\method{cjsAddSeries}{cjs_radar}(cjs, data, type = NULL, color = NULL,
  label = NULL, scale = NULL)

\method{cjsAddSeries}{cjs_pie}(cjs, data, color = NULL)

\method{cjsAddSeries}{cjs_polarArea}(cjs, data, color = NULL)
}
\arguments{
\item{cjs}{a chartjs object}

\item{data}{a vector of data}

\item{type}{a character or NULL, the representation of the data,}

\item{color}{an optional list of colours, see \link{colortypes}}

\item{label}{a character, the name of the series}

\item{scale}{an integer, the numeric position of the scale to attach the series to}
}
\description{
Use this function to add a series to a chart.
}
\section{Methods (by class)}{
\itemize{
\item \code{cjs_bar}: Add series to a bar chart

\item \code{cjs_line}: Add series to a line chart

\item \code{cjs_radar}: Add series to a radar chart

\item \code{cjs_pie}: Add series to a pie chart

\item \code{cjs_polarArea}: Add series to a polarArea chart
}}
\keyword{internal}

