% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjsBar.R
\name{cjsBar}
\alias{cjsBar}
\title{Bar chart}
\usage{
cjsBar(chartjs, stacked = FALSE)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{stacked}{a logical, if FALSE (the defaults), bars will be dodged,
else they will be stacked.}
}
\description{
This function takes a \code{chartjs} object created by the \code{\link{chartjs}}
function and draws a bar chart.
}
\details{
The \code{\link{chartjs}} object \code{data} parameter should be a named list
of numeric vectors.
}

