% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_scale.R
\name{cjsEditScale}
\alias{cjsEditScale}
\title{Edit an existing scale}
\usage{
cjsEditScale(cjs, axis = c("x", "y", NULL), scale, ...)
}
\arguments{
\item{cjs}{a \code{\link{chartjs}} object}

\item{axis}{the axis on which the scale lives}

\item{scale}{the numeric position of the scale}

\item{...}{parameters to edit}
}
\description{
\code{cjsEditScale} allows the editing of an existing scale.
}
\details{
See the \href{http://www.chartjs.org/docs/#scales}{Charts.js documentation} for values in ...
}

