% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{cjsInteraction}
\alias{cjsInteraction}
\title{Description of the cjsInteraction function.}
\usage{
cjsInteraction(chartjs, ...)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{...}{named arguments, see Details}
}
\description{
Description of the cjsInteraction function.
}
