% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{cjsLegend}
\alias{cjsLegend}
\title{Add legends}
\usage{
cjsLegend(chartjs, ...)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{...}{a list of additional parameters}
}
\description{
Add legends
}

