% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjsLine.R
\name{cjsLine}
\alias{cjsLine}
\title{Line chart}
\usage{
cjsLine(chartjs)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}
}
\description{
This function takes a \code{chartjs} object created by the \code{\link{chartjs}}
function and draws a line chart.
}
\details{
The \code{\link{chartjs}} object \code{data} parameter should be a named list
of numeric vectors.
}

