% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjsPie.R
\name{cjsPie}
\alias{cjsPie}
\title{Pie chart}
\usage{
cjsPie(chartjs)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}
}
\description{
This function takes a \code{chartjs} object created by the \code{\link{chartjs}}
function and draws a pie chart.
}
\details{
The \code{\link{chartjs}} object \code{data} parameter should be a named vector (or list)
of numerics
}

