% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{cjsScale}
\alias{cjsScale}
\title{Scales}
\usage{
cjsScale(chartjs, axis = c("x", "y"), type = NULL, display = NULL,
  callbacks = NULL, gridLines = NULL, scaleLabel = NULL, ticks = NULL,
  position = NULL, ...)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{axis}{a character, the axis to modify, either "x" or "y"}

\item{type}{a character, the type of scale for this axis}

\item{display}{a logical, should this scale be displayed}

\item{callbacks}{a list of Javascript functions as characters. Note that these are not bundled in
the Chart.js library, only in the R package for easier definition. Remember to wrap those with the
\code{\link[htmlwidgets]{JS}} function}

\item{gridLines}{a list of grid lines parameters}

\item{scaleLabel}{a list of parameters to modify the appearance of the labels}

\item{ticks}{a list of parameters to modify the appearance of the ticks}

\item{position}{a character}

\item{...}{additional parameters for the differents scales categories (linear, logarithmic, time, etc.)}
}
\description{
Modify an x- or y- scale. As of now, multiple scales on the same axis are not supported (but they
are in Chart.js). See the \href{http://nnnick.github.io/Chart.js/docs-v2/#scales}{Chart.js scales documentation}
for all the options (charts also have additional scales options).
}

