% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{cjsTitle}
\alias{cjsTitle}
\title{Add title}
\usage{
cjsTitle(chartjs, title, ...)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{title}{the title of the chart}

\item{...}{additional parameters}
}
\description{
Add title
}

