% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/colortypes.R
\name{colortypes}
\alias{colortypes}
\title{Color types}
\arguments{
\item{borderColor}{}

\item{backgroundColor}{}

\item{pointBorderColor}{}

\item{pointBackgroundColor}{}

\item{hoverBorderColor}{}

\item{hoverBackgroundColor}{}

\item{pointHoverBorderColor}{}

\item{pointHoverBackgroundColor}{}
}
\description{
Allowed color types
}

