% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/doughnutchart.R
\name{doughnutChart}
\alias{doughnutChart}
\title{Draw a doughnut chart}
\usage{
doughnutChart(chartjs, colours = NULL, inner = 50)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{colours}{an optional list of colours}

\item{inner}{a numeric, the percentage of the inner cutout}
}
\description{
Draw a doughnut chart
}

