% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{guess_label}
\alias{guess_label}
\title{Guess Chart Label}
\usage{
guess_label(x)
}
\arguments{
\item{x}{The input code (passed through \code{substitute} first) to deparse.}
}
\value{
A guess for an appropriate label to use (character scalar).
}
\description{
Deparse some input code to guess a good label for a chart.
}
