% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartjs.R
\name{is.chartjs}
\alias{is.chartjs}
\title{Check that an Object is a \code{Chart.js} Plot Object}
\usage{
is.chartjs(x)
}
\arguments{
\item{x}{The object to check (any R object).}
}
\value{
Whether the object inherits from \code{chartjs}.
}
\description{
Check that an Object is a \code{Chart.js} Plot Object
}
\examples{
is.chartjs(mtcars)                     # FALSE
is.chartjs(chartjs(mtcars, x = ~ mpg)) # TRUE

}
