% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{make_scale}
\alias{make_scale}
\title{Make a scale object in the format required for \code{Chart.js}}
\usage{
make_scale(
  id,
  pos = c("left", "right", "bottom", "top"),
  display = TRUE,
  grid = TRUE
)
}
\description{
Make a scale object in the format required for \code{Chart.js}
}
