% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{non_null}
\alias{non_null}
\title{Remove NULLs from a List}
\usage{
non_null(x)
}
\arguments{
\item{x}{The vector/list to remove NULL elements from.}
}
\value{
A filtered vector/list.
}
\description{
Remove NULLs from a List
}
\examples{
non_null(1:10)
non_null(list(1, "a", NULL, 10L))

}
