% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{cjsOptions}
\alias{options}
\title{Edit default options}
\usage{
cjsOptions(chartjs, ...)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{...}{named arguments, see Details}
}
\description{
Allows editing of the chart options.
}
\details{
The allowed values in \code{...} can be found in the
\href{http://www.chartjs.org/docs/#getting-started-global-chart-configuration}{Chart.js documentation}
under \code{Charts.default.global}. You can also use this function to edit the \code{animation} and \code{elements}
options by passing them as a list, or to edit the chart-specific options.
To edit tooltips, legends or titles, use the corresponding functions.
}

