% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/piechart.R
\name{pieChart}
\alias{pieChart}
\title{Draw a pie chart}
\usage{
pieChart(chartjs, colours = NULL, inner = 0)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{colours}{an optional list of colours}

\item{inner}{a numeric, the percentage of the inner cutout}
}
\description{
Draw a pie chart
}

