% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{a \code{\link{chartjs}} object}

\item{rhs}{a charting function}
}
\description{
Imports the pipe operator from magrittr.
}

