% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/chartjs.R
\name{renderChartjs}
\alias{renderChartjs}
\title{Widget render function for use in Shiny}
\usage{
renderChartjs(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

