% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_padding}
\alias{set_padding}
\title{Set Chart Padding}
\usage{
set_padding(
  chart,
  all = NULL,
  left = all,
  right = all,
  top = all,
  bottom = all
)
}
\arguments{
\item{chart}{A \code{chartjs} object.}

\item{all}{The padding to apply to each side (numeric scalar).}

\item{left, right, top, bottom}{The padding to apply to each individual side (overrides \code{all}) (numeric scalar).}
}
\value{
A modified \code{chartjs} object.
}
\description{
Set the padding around a \code{chartjs} object
}
\details{
See here for more detailed documentation:
  \url{https://www.chartjs.org/docs/latest/configuration/layout.html?h=padding}.
}
\examples{

}
