% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny-chartjs}
\alias{chartjsOutput}
\alias{renderChartjs}
\alias{shiny-chartjs}
\title{Shiny bindings for chartjs}
\usage{
chartjsOutput(outputId, width = "100\%", height = "400px")

renderChartjs(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{the widget id in the server part of the application}

\item{width}{the width of the widget}

\item{height}{the height of the widget}

\item{expr}{an R expression that evaluates to a \code{\link{chartjs}} object}

\item{env}{the environment in which to evaluate \code{expr}}

\item{quoted}{a boolean, is \code{expr} a quoted expression ?}
}
\description{
Shiny bindings for chartjs
}

