% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{alter_axis}
\alias{alter_axis}
\title{Alter an Axis}
\usage{
alter_axis(
  chart,
  id,
  xOrY = tolower(substr(id, 1, 1)),
  ...,
  pos,
  display = TRUE,
  grid = TRUE,
  type = "linear",
  title,
  label,
  percent = FALSE,
  digits = 2L,
  min,
  max,
  suggestedMin,
  suggestedMax
)
}
\arguments{
\item{pos}{One of: c("left", "right", "bottom", "top")}
}
\description{
Alter an Axis
}
