% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basechart.R
\docType{data}
\name{baseChart}
\alias{.vecColors}
\alias{baseChart}
\title{Base chart creation}
\format{An object of class \code{character} of length 12.}
\usage{
.vecColors
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{type}{a character, the type of chart}

\item{colours}{an optional list of colours. If NULL, default colors from the
\strong{Paired} palette of the RColorBrewer package will be used.}
}
\description{
This is a low-level function called by the charts creation functions. It adds
colors and chart type to the object that will be interpreted by the Chart.js
library.
}
\keyword{internal}

