% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/basechart.R
\docType{data}
\name{baseChart}
\alias{.vecColors}
\alias{baseChart}
\title{Base chart creation}
\format{\preformatted{ chr [1:12] "#A6CEE3" "#1F78B4" "#B2DF8A" "#33A02C" "#FB9A99" "#E31A1C" "#FDBF6F" "#FF7F00" "#CAB2D6" ...
}}
\usage{
.vecColors
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{type}{a character, the type of chart}

\item{colours}{an optional list of colours. If NULL, default colors from the
\strong{Paired} palette of the RColorBrewer package will be used.}
}
\description{
This is a low-level function called by the charts creation functions. It adds
colors and chart type to the object that will be interpreted by the Chart.js
library.
}
\keyword{internal}

