% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartjs.R
\name{chartjs}
\alias{chartjs}
\title{Create a \code{Chart.js} Plot}
\usage{
chartjs(
  data,
  x,
  type = "bar",
  options = baseOptions,
  forceBig = FALSE,
  width = "100\%",
  height = "100\%",
  elementId = NULL
)
}
\arguments{
\item{data}{}

\item{x}{}

\item{type}{}

\item{options}{}

\item{width}{}

\item{height}{}

\item{elementId}{}
}
\value{

}
\description{
Create a \code{Chart.js} Plot
}
\examples{
chartjs(mtcars, x = ~ mpg) \%>\% new_bars(y = ~ disp)

}
