% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartjs.R
\name{chartjs}
\alias{chartjs}
\title{chartjs}
\usage{
chartjs(data, ..., labels = NULL, dataLabels = NULL, width = NULL,
  height = NULL, chartOptions = NULL)
}
\arguments{
\item{data}{a data.frame,}

\item{...}{the columns of \code{data} to plot,}

\item{labels}{a vector of numerics or characters : the labels on the x-axis,}

\item{dataLabels}{a vector of characters : the names of the data series,
defaults to the names of \code{data}}

\item{width}{the width of the widget,}

\item{height}{the height of the widget,}

\item{chartOptions}{a list of options to customize the graph,
see the \href{http://nnnick.github.io/Chart.js/docs-v2/}{Chart.js documentation}.}
}
\description{
Draw a graph using the Chart.js library.
}
\details{
See the \href{http://nnnick.github.io/Chart.js/docs-v2/}{Chart.js documentation}.
}

