% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartjs.R
\name{chartjs}
\alias{chartjs}
\title{chartjs}
\usage{
chartjs(width = NULL, height = NULL,
  palette = RColorBrewer::brewer.pal(12, "Paired"), debug = FALSE)
}
\arguments{
\item{width}{the width of the widget}

\item{height}{the height of the widget}

\item{palette}{a color palette}

\item{debug}{if TRUE, prints information to the JS console}
}
\description{
Draw a graph using the Chart.js library.
}

