% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chartjs.R
\name{chartjs}
\alias{chartjs}
\title{chartjs}
\usage{
chartjs(data, labels = NULL, width = NULL, height = NULL,
  chartOptions = NULL)
}
\arguments{
\item{data}{a named list of the y-axis data}

\item{labels}{a vector of numerics or characters : the labels of the data points.
Not needed for Pie, Doughnut and PolarArea charts.}

\item{width}{the width of the widget}

\item{height}{the height of the widget}

\item{chartOptions}{a list of chartOptions to customize the graph}
}
\description{
Draw a graph using the Chart.js library
}
\details{
See the documentation of Chart.js at \url{http://www.chartjs.org/docs}.
}

