% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/chartjs.R
\name{chartjs}
\alias{chartjs}
\alias{chartjs_}
\title{chartjs}
\usage{
chartjs(data, ..., labels = NULL, dataLabels = NULL, title = NULL,
  width = NULL, height = NULL, chartOptions = NULL)

chartjs_(data, labels = NULL, dataLabels = NULL, title = NULL,
  width = NULL, height = NULL, chartOptions = NULL)
}
\arguments{
\item{data}{a dataframe;}

\item{...}{unquoted names of the columns to plot,}

\item{labels}{a vector of numerics or characters : the labels on the x-axis.
Not needed for Pie, Doughnut and PolarArea charts,}

\item{dataLabels}{a vector of characters : the names of the data series,}

\item{title}{the title of the chart,}

\item{width}{the width of the widget,}

\item{height}{the height of the widget,}

\item{chartOptions}{a list of options to customize the graph, see the Chart.js documentation.}

\item{dataframe}{a dataframe,}
}
\description{
Draw a graph using the Chart.js library. \code{chartjs} uses NSE to draw specific columns from a data.frame
while \code{chartjs_} draws all columns from a data.frame.
}
\details{
See the documentation of Chart.js at \url{http://www.chartjs.org/docs}.
}

