% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartjs.R
\name{chartjs}
\alias{chartjs}
\title{chartjs}
\usage{
chartjs(width = NULL, height = NULL, palette = "Paired", debug = FALSE)
}
\arguments{
\item{width}{the width of the widget}

\item{height}{the height of the widget}

\item{palette}{either the name of a RColorBrewer palette or a list of custom colours as specified in \code{\link{colorTypes}}}

\item{debug}{if TRUE, prints information to the JS console}
}
\description{
Draw a graph using the Chart.js library.
}

