% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-proxy.R
\name{chartjs_add_dataset}
\alias{chartjs_add_dataset}
\title{Add a new dataset to the chart}
\usage{
chartjs_add_dataset(proxy, dataset)
}
\arguments{
\item{proxy}{A \code{chartjs_proxy} object created with \link{chartjs_proxy}}

\item{dataset}{A list describing the new dataset. This list should follow the
Chart.js dataset structure (e.g. contain \code{label}, \code{data}, \code{backgroundColor}).}
}
\value{
The proxy object (for method chaining)
}
\description{
Add a new dataset to the chart
}
