% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-proxy.R
\name{chartjs_update_options}
\alias{chartjs_update_options}
\title{Update chart options via proxy}
\usage{
chartjs_update_options(proxy, options)
}
\arguments{
\item{proxy}{A \code{chartjs_proxy} object created with \link{chartjs_proxy}}

\item{options}{New options for the chart. These will be merged with the
current chart options on the client.}
}
\value{
The proxy object (for method chaining)
}
\description{
Update chart options via proxy
}
