% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartjs.R
\name{chartjsOutput}
\alias{chartjsOutput}
\title{Widget output function for use in Shiny}
\usage{
chartjsOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{height}{Same as width}
}
\description{
Widget output function for use in Shiny
}
