% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/charts.R
\name{charts}
\alias{barChart}
\alias{charts}
\alias{doughnutChart}
\alias{lineChart}
\alias{pieChart}
\alias{polarAreaChart}
\alias{radarChart}
\alias{scatterChart}
\title{Draw charts}
\usage{
barChart(chartjs, colours = NULL, stacked = FALSE)

lineChart(chartjs, colours = NULL)

pieChart(chartjs, colours = NULL, inner = 0)

doughnutChart(chartjs, colours = NULL, inner = 50)

radarChart(chartjs, colours = NULL)

polarAreaChart(chartjs, colours = NULL)

scatterChart(chartjs, colours = NULL)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{colours}{an optional list of colours.}

\item{stacked}{a logical, defaults to FALSE. If TRUE, bars will be stacked
at each x value.}

\item{inner}{the optional percentage of the inner cutout for Pie and Doughnut
charts (defaults to 0 for Pie and 50 for Doughnut)}
}
\description{
These functions are used to draw the different types of graphs.
}
\details{
For \code{colours}, the names of the list must be among those in \code{\link{colortypes}}.
The default colours are obtained from the \strong{Paired} palette from the
\code{\link[RColorBrewer]{RColorBrewer}} package.
}
\examples{
labels <- row.names(mtcars)
data <- list(mpg = mtcars$mpg, qsec = mtcars$qsec)
dataPie <- mtcars \%>\%
  dplyr::count(cyl)
dataScatter <- list(y = matrix(round(runif(20),2), ncol = 2), z =matrix(round(runif(20),2), ncol = 2))

# Draw a bar chart
chartjs(data, labels) \%>\%
  barChart

# Draw a stacked bar chart
chartjs(data, labels) \%>\%
  barChart(stacked = TRUE)

# Draw a line chart
chartjs(data, labels) \%>\%
  lineChart

# Draw a radar chart
chartjs(data, labels) \%>\%
  radarChart

# Draw a pie chart
chartjs(as.list(dataPie$n), labels = dataPie$cyl) \%>\%
  pieChart

# Draw a doughnut chart
chartjs(as.list(dataPie$n), labels = dataPie$cyl) \%>\%
  doughnutChart

# Draw a scatter chart
chartjs(dataScatter) \%>\%
  scatterChart
}

