% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/charts.R
\name{charts}
\alias{barChart}
\alias{charts}
\alias{doughnutChart}
\alias{lineChart}
\alias{pieChart}
\alias{polarAreaChart}
\alias{radarChart}
\title{Draw charts}
\usage{
barChart(chartjs, colours = NULL)

lineChart(chartjs, colours = NULL)

pieChart(chartjs, colours = NULL, inner = 0)

doughnutChart(chartjs, colours = NULL, inner = 50)

radarChart(chartjs, colours = NULL)

polarAreaChart(chartjs, colours = NULL)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{colours}{an optional list of colours.}

\item{inner}{the optional percentage of the inner cutout for Pie and Doughnut charts (default 0 for Pie and 50 for Doughnuts)}
}
\description{
These functions are used to draw the different types of graphs.
}
\details{
For \code{colours}, the names of the list must be among those in \code{\link{colortypes}}.
The base colours are obtained from the \strong{Paired} palette from the \code{\link[RColorBrewer]{RColorBrewer}} package.
}

