% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{charts}
\alias{charts}
\alias{cjsBar}
\alias{cjsDoughnut}
\alias{cjsLine}
\alias{cjsPie}
\alias{cjsPolar}
\alias{cjsRadar}
\title{Draw charts}
\usage{
cjsBar(chartjs, labels, ...)

cjsLine(chartjs, labels, ...)

cjsPie(chartjs, labels, ...)

cjsDoughnut(chartjs, cutout = 50)

cjsPolar(chartjs)

cjsRadar(chartjs)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{cutout}{the optional percentage of the inner cutout for Doughnut
charts (defaults to 50)}

\item{stacked}{a logical, defaults to FALSE. If TRUE, bars will be stacked
at each x value.}
}
\description{
These functions are used to draw the different types of graphs.
}
\details{
For \code{colours}, the names of the list must be among those in \code{\link{colortypes}}.
The default colours are obtained from the \strong{Paired} palette from the
\code{\link[RColorBrewer]{RColorBrewer}} package.
}
\examples{
# Draw a bar chart
chartjs(mtcars, mpg, qsec) \%>\%
  cjsBar

# Draw a stacked bar chart
chartjs(mtcars, mpg, qsec) \%>\%
  cjsBar(stacked = TRUE)

# Draw a line chart
chartjs(mtcars, mpg, qsec) \%>\%
  cjsLine

# Draw a radar chart
chartjs(mtcars, mpg, qsec) \%>\%
  cjsRadar

# Draw a pie chart
chartjs(mtcars[1:6,], mpg) \%>\%
  cjsPie

# Draw a doughnut chart
chartjs(mtcars[1:6,], mpg) \%>\%
  cjsDoughnut(cutout = 80)

# Draw a polar area chart
chartjs(mtcars[1:6,], mpg) \%>\%
  cjsPolar

}

