% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_scale.R
\name{cjsAddScale}
\alias{cjsAddScale}
\title{Add a scale to an axis}
\usage{
cjsAddScale(cjs, axis = c("x", "y", NULL), type = "category", ...)
}
\arguments{
\item{cjs}{a \code{\link{chartjs}} object}

\item{axis}{the axis on which the scale lives}

\item{type}{the type of scale to add}

\item{...}{additional parameters}
}
\description{
\code{cjsAddScale} adds a scale to an axis.
}
\details{
See the \href{http://www.chartjs.org/docs/#scales}{Charts.js documentation} for values in ...
}

