% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{cjsLegend}
\alias{cjsLegend}
\title{Edit legend}
\usage{
cjsLegend(chartjs, position = "right", title = "", template = NULL)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{position}{a character to specify if the legend is on the right or on the bottom of the chart. Not implemented yet.}

\item{title}{a character, the title of the legend.}

\item{template}{a Javascript template for the legend. If NULL (the default), the standard Chart.js template is used}

\item{direction}{a character to specify if the legend extends horizontally or vertically. Not implemented yet.}
}
\description{
Edit legend
}
\examples{
chartjs(data, labels, title = "mpg and qsec in the mtcars dataset") \%>\%
  barChart \%>\%
  cjsLegend(title = "Variables")
}

