% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{cjsLegend}
\alias{cjsLegend}
\title{Add legends}
\usage{
cjsLegend(chartjs, position = NULL, onClick = NULL, fullWidth = NULL,
  labels = NULL)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{position}{"top" or "bottom", defaults to "top"}

\item{onClick}{a JS function as a character}

\item{fullWidth}{a logical, defaults to TRUE}

\item{labels}{a list of parameters for the labels : \itemize{
 \item boxWidth an integer, defaults to 40
 \item fontColor a character, defaults to "#666"
 \item fontFamily a character, defaults to "Helvetica Neue"
 \item fontSize an integer, defaults to 12
 \item fontStyle a character, defaults to "normal"
 \item padding a integer, defaults to 10 (in pixels)
}}
}
\description{
Add legends
}

