% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjsPie.R
\name{cjsPie}
\alias{cjsDoughnut}
\alias{cjsPie}
\title{Pie and doughnut charts}
\usage{
cjsPie(chartjs)

cjsDoughnut(chartjs, cutout = 50)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{cutout}{a numeric, the percentage of the chart that is cut out of the middle}
}
\description{
These functions take a \code{chartjs} object created by the \code{\link{chartjs}}
function and draw a pie or a doughnut chart.
}
\details{
The \code{\link{chartjs}} object \code{data} parameter should be a named vector (or list)
of numerics
}

