% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{cjsSeries}
\alias{cjsSeries}
\alias{cjsSeries.cjs_bar}
\alias{cjsSeries.cjs_horizontalBar}
\alias{cjsSeries.cjs_line}
\alias{cjsSeries.cjs_pie}
\alias{cjsSeries.cjs_polarArea}
\alias{cjsSeries.cjs_radar}
\title{Add a series to a chart}
\usage{
cjsSeries(cjs, data, type, label, scale)

\method{cjsSeries}{cjs_bar}(cjs, data, type = "bar", label = NULL,
  scale = NULL)

\method{cjsSeries}{cjs_line}(cjs, data, type = "line", label = NULL,
  scale = NULL)

\method{cjsSeries}{cjs_horizontalBar}(cjs, data, type = "horizontalBar",
  label = NULL, scale = NULL)

\method{cjsSeries}{cjs_radar}(cjs, data, type = NULL, label = NULL,
  scale = NULL)

\method{cjsSeries}{cjs_pie}(cjs, data, type = NULL, label = NULL,
  scale = NULL)

\method{cjsSeries}{cjs_polarArea}(cjs, data, type = NULL, label = NULL,
  scale = NULL)
}
\arguments{
\item{cjs}{a \code{\link{chartjs}} object}

\item{data}{a vector of data}

\item{type}{a character or NULL, the representation of the data,}

\item{label}{a character, the name of the series}

\item{scale}{an integer, the numeric position of the scale to attach the series to}

\item{color}{an optional list of colours, see \link{colortypes}}

\item{palette}{an optional RColorBrewer palette, defaults to Paired, see \code{\link[RColorBrewer]{display.brewer.all}}}
}
\description{
Use this function to add a series to a chart.
}
\section{Methods (by class)}{
\itemize{
\item \code{cjs_bar}: Add series to a bar chart

\item \code{cjs_line}: Add series to a line chart

\item \code{cjs_horizontalBar}: Add series to a horizontal bar chart

\item \code{cjs_radar}: Add series to a radar chart

\item \code{cjs_pie}: Add series to a pie chart

\item \code{cjs_polarArea}: Add series to a polarArea chart
}}
\keyword{internal}

