% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{cjsTitle}
\alias{cjsTitle}
\title{Add title}
\usage{
cjsTitle(chartjs, text)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{text}{a character}

\item{direction}{a character to specify if the legend extends horizontally or vertically. Not implemented yet.}

\item{title}{a character, the title of the legend.}

\item{template}{a Javascript template for the legend. If NULL (the default), the standard Chart.js template is used}
}
\description{
Add title
}

