% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/colortypes.R
\name{colortypes}
\alias{colortypes}
\title{Color types}
\arguments{
\item{strokeColor}{}

\item{fillColor}{}

\item{highlightStroke}{}

\item{highlightFill}{}

\item{pointColor}{}

\item{pointStrokeColor}{}

\item{pointHighlightFill}{}

\item{pointHighlightStroke}{}

\item{color}{}

\item{highlight}{}
}
\description{
Allowed color types
}

