% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{colortypes}
\alias{colortypes}
\title{Color types}
\arguments{
\item{borderColor}{for all charts}

\item{backgroundColor}{for all charts}

\item{pointBorderColor}{for Line, Radar and Scatter charts}

\item{pointBackgroundColor}{for Line, Radar and Scatter charts}

\item{hoverBorderColor}{for Bar, Pie, Doughnut and PolarArea charts}

\item{hoverBackgroundColor}{for Bar, Pie, Doughnut and PolarArea charts}

\item{pointHoverBorderColor}{for Line, Radar and Scatter charts}

\item{pointHoverBackgroundColor}{for Line, Radar and Scatter charts}
}
\description{
Allowed color types
}
\details{
When adding custom colours to a chart, the color argument should have the structure displayed in
the example and \strong{ALL} necessary colors must be given.
}
\examples{
color <- list(background = "#FF0000",
              border = "#00FF00",
              hover = list(background = "#FF0000",
                           border = "#00FF00"),
              point = list(background = "#FF0000",
                           border = "#00FF00",
                           hover = list(background = "#FF0000",
                                        border = "#00FF00")))
}

