% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{make_dataset}
\alias{make_dataset}
\title{Make a dataset in the format required for \code{Chart.js}}
\usage{
make_dataset(
  data,
  y,
  label = NULL,
  type,
  ...,
  yAxis = NULL,
  xAxis = NULL,
  order = NULL,
  legend = NULL,
  fill = (type != "line"),
  bgCol = NULL,
  brdCol = bgCol,
  brdWidth = NULL,
  radius = NULL,
  hoverRadius = NULL,
  tension = NULL,
  stepped = list(NULL, FALSE, TRUE, "before", "after", "middle")[[1]],
  barPerc = NULL,
  catPerc = NULL
)
}
\description{
Make a dataset in the format required for \code{Chart.js}
}
