% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltips.R
\name{tooltips}
\alias{cjsTooltips}
\alias{tooltips}
\title{Edit default tooltips}
\usage{
cjsTooltips(chartjs, ...)
}
\description{
Allows editing of the default tooltips.
}
\details{
The allowed values in \code{...} are :

\tabular{ll}{
  enabled \tab logical, defaults to TRUE \cr
  custom  \tab character containing a Javascript function, defaults to 'null' \cr
  backgroundColor \tab character, defaults to 'rgba(0,0,0,0.8)' \cr
  fontFamily \tab character vector, defaults to "'Helvetica Neue', 'Helvetica', 'Arial', sans-serif" \cr
  fontSize \tab numeric, defaults to 10 \cr
  fontStyle \tab character, defaults to "normal" \cr
  fontColor \tab character, defaults to "#fff" \cr
  titleFontFamily \tab character, defaults to "'Helvetica Neue', 'Helvetica', 'Arial', sans-serif" \cr
  titleFontSize \tab numeric, defaults to 12 \cr
  titleFontStyle \tab character, defaults to "bold" \cr
  titleColor \tab character, defaults to "#fff" \cr
  yPadding \tab numeric, defaults to 6 \cr
  xPadding \tab numeric, defaults to 6 \cr
  caretSize \tab numeric, defaults to 8 \cr
  cornerRadius \tab numeric, defaults to 6 \cr
  xOffset \tab numeric, defaults to 10 \cr
  template \tab character, defaults to "<\% if(label){ \%><\%=label \%>: <\% } \%><\%=value \%>" \cr
  multiTemplate \tab character, defaults to "<\%if (datasetLabel){ \%><\%=datasetLabel \%>: <\% } \%><\%=value \%>" \cr
  multiKeyBackground \tab character, defaults to '#fff' \cr
}
}
\examples{
labels <- row.names(mtcars)
data <- list(mpg = mtcars$mpg, qsec = mtcars$qsec)

# Change tooltip title color
chartjs(data, labels) \%>\%
  cjsBar \%>\%
  cjsTooltips(titleFontColor = "#f0f")
}

