% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltips.R
\name{tooltips}
\alias{cjsTooltips}
\alias{tooltips}
\title{Edit default tooltips}
\usage{
cjsTooltips(chartjs, ...)
}
\arguments{
\item{chartjs}{a \code{\link{chartjs}} object}

\item{...}{named arguments, see Details}
}
\description{
Allows editing of the default tooltips.
}
\details{
The allowed values in \code{...} can be found in the
\href{http://www.chartjs.org/docs/#getting-started-global-chart-configuration}{Chart.js documentation}
under \code{Charts.default.global.tooltips}.
}

